moving file button to shared UI
