+++ inspector/src/components/actionTabs/lines/textureLinkLineComponent.tsx
@@ -10,7 +10,7 @@ import { TextLineComponent } from "./textLineComponent";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faWrench, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { Texture } from 'babylonjs/Materials/Textures/texture';
-import { FileButtonLineComponent } from './fileButtonLineComponent';
+import { FileButtonLineComponent } from './../../../sharedUiComponents/lines/fileButtonLineComponent';
 import { Tools } from 'babylonjs/Misc/tools';
 
 export interface ITextureLinkLineComponentProps {
